+++ src/main/java/com/sparta/blog/commonDto/CommonResponseDto.java
@@ -1,4 +1,4 @@
-package com.sparta.blog.dto;
+package com.sparta.blog.commonDto;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
 import lombok.AllArgsConstructor;
